requirements.txt
opencv-python==4.5.1.48; platform_machine != "aarch64" and platform_machine != "armv6l" and platform_machine != "armv7l"
opencv-contrib-python==4.5.1.48; platform_machine != "aarch64" and platform_machine != "armv6l" and platform_machine != "armv7l"
opencv-python==4.4.0.46; platform_machine == "armv6l" or platform_machine == "armv7l"
opencv-contrib-python==4.4.0.46; platform_machine == "armv6l" or platform_machine == "armv7l"
--extra-index-url https://artifacts.luxonis.com/artifactory/luxonis-python-snapshot-local/
depthai==2.10.0.0.dev+076c968eb8b0439a53aecb508511fbf932d8eddd
imutils==0.5.4
pygame==2.0.1
pygame-gui==0.5.7
PyOpenGL==3.1.5
PyWavefront==1.3.3